Project_Webapp/django Integration/requirements.txt
Django>=3.1,<4.0
djangorestframework>=3.11.1
numpy
scikit-learn
django-cors-headers
beautifulsoup4
python-whois
requests
pandas
matplotlib
seaborn